Move plugins/selection.py to _internal/
