+++ certbot/tests/main_test.py
@@ -1220,7 +1220,7 @@ class MainTest(test_util.ConfigTestCase):  # pylint: disable=too-many-public-met
         self._test_renew_common(renewalparams=renewalparams, error_expected=True,
                                 names=names, assert_oc_called=False)
 
-    @mock.patch('certbot.plugins.selection.choose_configurator_plugins')
+    @mock.patch('certbot._internal.plugins.selection.choose_configurator_plugins')
     def test_renew_with_configurator(self, mock_sel):
         mock_sel.return_value = (mock.MagicMock(), mock.MagicMock())
         renewalparams = {'authenticator': 'webroot'}
@@ -1493,7 +1493,7 @@ class MainTest(test_util.ConfigTestCase):  # pylint: disable=too-many-public-met
                             email in mock_utility().add_message.call_args[0][0])
                         self.assertTrue(mock_handle.called)
 
-    @mock.patch('certbot.plugins.selection.choose_configurator_plugins')
+    @mock.patch('certbot._internal.plugins.selection.choose_configurator_plugins')
     @mock.patch('certbot._internal.updater._run_updaters')
     def test_plugin_selection_error(self, mock_run, mock_choose):
         mock_choose.side_effect = errors.PluginSelectionError
